Created the PostAddress usecases and refactored their location
