--- code/orders-domain/src/main/java/com/academy/orders/domain/orderV2/repository/PostAddressRepository.java
@@ -1,4 +1,0 @@
-package com.academy.orders.domain.orderV2.repository;
-
-public interface PostAddressRepository {
-}
