+++ code/orders-domain/src/main/java/com/academy/orders/domain/postaddress/repository/PostAddressRepository.java
@@ -1,0 +1,4 @@
+package com.academy.orders.domain.postaddress.repository;
+
+public interface PostAddressRepository {
+}
